development

+++ src/modules/users/application/delete_by_id/use_case.py
@@ -1,5 +1,6 @@
 from src.common.use_case import UseCase
 
+from ...exceptions.user_errors import UserDoesNotExistError
 from ...infrastructure.repositories.user_unit_of_work import UserUnitOfWork
 
 
@@ -8,6 +9,11 @@ class DeleteUserByIdUseCase(UseCase[str, str]):
         self.unit_of_work = unit_of_work
 
     def __call__(self, user_id: str) -> str:
+        existing_user = self.unit_of_work.repository.find_by_id(user_id)
+
+        if not existing_user:
+            raise UserDoesNotExistError()
+
         result = self.unit_of_work.repository.delete_by_id(user_id)
         self.unit_of_work.commit()
         return result
